requirements.txt
Flask
flask-cors
requests
google-generativeai
beautifulsoup4
selenium
webdriver-manager
firebase-admin
gTTS
uuid